// src/types/data.ts
export interface Product {
  id: number
  name: string
  price: number
  oldPrice?: number
  description: string
  category: string
  inStock: boolean
  images: string[]
  features: string[]
  specifications: {
    brand?: string
    collection?: string
    style?: string
    room?: string
    warranty?: string
    country?: string
  }
}